% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.mf.R
\name{mice.impute.mf}
\alias{mice.impute.mf}
\title{Single imputation using random forest algorithm}
\usage{
mice.impute.mf(
  y,
  ry,
  x,
  wy = NULL,
  num.trees = 10,
  num.trees.continuous = NULL,
  num.trees.categorical = NULL,
  use.pred.prob.categorical = FALSE,
  pre.bootstrap = FALSE,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees}{Number of trees to build, default to \code{10}, overrides
\code{num.trees.continuous} and \code{num.trees.categorical}}

\item{num.trees.continuous}{Number of trees to build for imputing continuous
variables, default to \code{10}}

\item{num.trees.categorical}{Number of trees to build for imputing
categorical variables, default to \code{10}}

\item{use.pred.prob.categorical}{Logical indicator, \code{TRUE} for
imputation based on probabilities of votes, \code{FALSE} for imputation
based on majority votes, default to \code{TRUE}}

\item{pre.bootstrap}{Perform bootstrap prior to imputation to get 'proper'
imputation, i.e. accommodating sampling variation in estimating population
regression parameters (see Shah et al. 2014)}

\item{...}{Other arguments to pass down}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Single imputation using random forest algorithm
}
\examples{
imp <- mice(nhanes, method = "mf")

}
\references{
Stekhoven, Daniel J., and Peter Bühlmann.
"MissForest—non-parametric missing value imputation for mixed-type data."
Bioinformatics 28.1 (2012): 112-118.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
