% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pkg-package.R
\docType{package}
\name{MiceMf-package}
\alias{MiceMf}
\alias{MiceMf-package}
\title{MiceMf: Single Imputation using MICE and random forest}
\description{
Adapter for MICE.
    Tools for performing single imputation using random forest algorithm.
}
\author{
\strong{Maintainer}: Shangzhi Hong \email{shangzhi-hong@hotmail.com} (\href{https://orcid.org/0000-0003-3843-7040}{ORCID})

Authors:
\itemize{
  \item Shangzhi Hong \email{shangzhi-hong@hotmail.com} (\href{https://orcid.org/0000-0003-3843-7040}{ORCID})
  \item Henry S. Lynn \email{hslynn@shmu.edu.cn} [thesis advisor]
}

}
\keyword{internal}
