% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiceMfNorm.R
\name{MiceMfNorm}
\alias{MiceMfNorm}
\title{MiceMfNorm: Single imputation using random forest algorithm}
\usage{
MiceMfNorm(data, num.trees = 10, maxit = 5, ...)
}
\arguments{
\item{data}{Data set containing missing values}

\item{num.trees}{Number of trees to build, default to \code{10}}

\item{maxit}{Number of iterations, default to \code{5}}
}
\value{
Imputed data set
}
\description{
MiceMfNorm: Single imputation using random forest algorithm
}
\examples{
impDf <- MiceMfNorm(nhanes)
}
