% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missForestMice.R
\name{missForestMice}
\alias{missForestMice}
\title{MissForest-flavored missing data imputation under MICE framework}
\usage{
missForestMice(
  xmis,
  maxiter = 5,
  ntree = 10,
  mtry = floor(sqrt(ncol(xmis))),
  decreasing = FALSE,
  ...
)
}
\arguments{
\item{xmis}{Data set containing missing values}

\item{maxiter}{Number of imputation iterations}

\item{ntree}{Number of trees to build}

\item{mtry}{Number of variables randomly sampled at each split}

\item{decreasing}{If \code{FALSE} then the variables are sorted according
increasing amount of missing cells.}

\item{...}{Other arguments to pass down}
}
\value{
Imputed complete dataset
}
\description{
MissForest-flavored missing data imputation under MICE framework
}
\examples{
imp <- missForestMice(nhanes)

}
\references{
Stekhoven, Daniel J., and Peter Bühlmann.
"MissForest—non-parametric missing value imputation for mixed-type data."
Bioinformatics 28.1 (2012): 112-118.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
